chore(astro): refactor actions
